// frontend/src/api/generated/PatchDecisionRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type PatchDecisionRequest = { kind: string | null, project_revision_id: string | null, basis: string | null, legal_reference: string | null, valid_from?: string | null, valid_to?: string | null, permitted_activities: Array<string> | null, conditions: Array<string> | null, restrictions: Array<string> | null, supersedes_id?: string | null, change_request_id?: string | null, document_version_id?: string | null, };